server/requirements.txt
Django==2.2.1
psycopg2-binary==2.8.2
dj-database-url==0.5.0
djangorestframework==3.10.3
django-cors-headers==3.2.0
django-rest-auth
django-allauth
gunicorn
factory-boy
Faker
